Refactored the relate process to be more uniform and remove all references to table
All references to ‘relate table’ have been removed since this is the only way we relate. Not used and untested code was removed so test coverage went up as well.
