+++ src/requirements.txt
@@ -8,7 +8,7 @@ coverage==4.5.1
 flake8==3.5.0
 GeoAlchemy2==0.5.0
 geomet==0.2.0.post2
--e git+https://github.com/Amsterdam/GOB-Core.git@v0.16.5#egg=gobcore
+-e git+https://github.com/Amsterdam/GOB-Core.git@v0.16.6#egg=gobcore
 idna==2.7
 Mako==1.0.7
 MarkupSafe==1.1.0
