+++ src/test.sh
@@ -13,4 +13,4 @@ echo "Running unit tests"
 pytest tests/
 
 echo "Running coverage tests"
-pytest --cov=gobupload --cov-report html --cov-fail-under=90
+pytest --cov=gobupload --cov-report html --cov-fail-under=91
